requirements.txt
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
fastapi==0.95.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
outcome==1.3.0.post0
psycopg2-binary==2.9.5
pycparser==2.22
pydantic==1.10.21
PySocks==1.7.1
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.9
starlette==0.26.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.21.1
websocket-client==1.8.0
wsproto==1.2.0
